move fn `create_merge_node` to mod `from_proto`
